refactor: endpointEntry
